+++ TrollsVsElves/TrollsVsElves/Core/Abstractions/ISingleton.cs
@@ -1,0 +1,4 @@
+﻿namespace TrollsVsElves.Core.Abstractions;
+
+public interface ISingleton
+{ }
