+++ TrollsVsElves/TrollsVsElves/Core/Abstractions/ITransient.cs
@@ -1,0 +1,4 @@
+﻿namespace TrollsVsElves.Core.Abstractions;
+
+public interface ITransient
+{ }
