--- TrollsVsElves/TrollsVsElves/Core/Lifetime/ISingleton.cs
@@ -1,4 +1,0 @@
-﻿namespace TrollsVsElves.Core.Lifetime;
-
-public interface ISingleton
-{ }
