--- TrollsVsElves/TrollsVsElves/Core/Lifetime/ITransient.cs
@@ -1,4 +1,0 @@
-﻿namespace TrollsVsElves.Core.Lifetime;
-
-public interface ITransient
-{ }
