--- TrollsVsElves/TrollsVsElves/Core/Textures/Transform.cs
@@ -1,28 +1,0 @@
-﻿using TrollsVsElves.Core.Lifetime;
-
-namespace TrollsVsElves.Core.Textures
-{
-    public class Transform : ITransient
-    {
-        private Vector2 _position;
-        private float _rotation;
-
-        private Vector2 _scale;
-
-        public Transform()
-        {
-            _scale = Vector2.One;
-            _position = Vector2.Zero;
-        }
-
-        public Vector2 Position => _position;
-        public float Rotation => _rotation;
-
-        public Vector2 Scale => _scale;
-
-        public void Translate(Vector2 translation)
-        {
-            _position += translation;
-        }
-    }
-}
