+++ TrollsVsElves/TrollsVsElves/Scripts/Scene.cs
@@ -1,8 +1,9 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Xna.Framework.Graphics;
 using Microsoft.Xna.Framework.Input;
+using TrollsVsElves.Core;
 using TrollsVsElves.Core.Extensions;
-using TrollsVsElves.Core.GameObjects;
+using TrollsVsElves.Core.Services;
 using TrollsVsElves.Core.Textures;
 using TrollsVsElves.Scripts.GameObjects;
 using TrollsVsElves.Scripts.Services;
@@ -39,15 +40,13 @@ public class Scene : Game
 
     protected override void Initialize()
     {
-        Core.GameWindow.Update(_graphics.PreferredBackBufferWidth, _graphics.PreferredBackBufferHeight);
+        GameWindowData.Update(_graphics.PreferredBackBufferWidth, _graphics.PreferredBackBufferHeight);
 
         _serviceCollection
             .AddSingleton<SpriteBatch>()
-            .AddSingleton<TextureFactory>()
-            .AddSingleton<InputHandlerService>()
-            .AddSingleton<GameObjectCollection>()
             .AddSingleton(Content)
-            .AddSingleton(GraphicsDevice);
+            .AddSingleton(GraphicsDevice)
+            .AddSingletonServices();
 
         _serviceCollection
             .AddTransientServices();
@@ -69,6 +68,7 @@ public class Scene : Game
         var player = provider.GetRequiredService<Player>();
 
         _gameObjectCollection.AddGameObject(player);
+
         _spriteBatch = provider.GetRequiredService<SpriteBatch>();
         _inputHandler = provider.GetRequiredService<InputHandlerService>();
     }
